chore: switch to interface and add type check test
